+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Both colorscad and 3mfmerge changes are included here. Unless explicitly mention
 
 ## [Unreleased]
 
+## [0.5.1] - 2021-08-22
+
 ### Fixed
 
 - Remove dependency on GNU coreutils on macOS; now OpenSCAD is the only dependency, no more need to brew
@@ -108,7 +110,8 @@ Both colorscad and 3mfmerge changes are included here. Unless explicitly mention
 
 - Everything! Only supports creating a colored .amf file.
 
-[Unreleased]: https://github.com/jschobben/colorscad/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/jschobben/colorscad/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/jschobben/colorscad/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/jschobben/colorscad/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/jschobben/colorscad/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/jschobben/colorscad/compare/v0.4.0...v0.4.1